# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Deltaswap Phylax API\nThis is the API for the Deltaswap Phylax and Explorer.\nThe API has two namespaces: deltaswapscan and phylax.\ndeltaswapscan is the namespace for the explorer and the new endpoints. The prefix is /api/v1.\nphylax is the legacy namespace backguard compatible with phylax node API. The prefix is /v1.\nThis API is public and does not require authentication although some endpoints are rate limited.\nCheck each endpoint documentation for more information.",
        "title": "Deltaswapscan API",
        "termsOfService": "https://wormhole.com/",
        "contact": {
            "name": "API Support",
            "url": "https://discord.com/invite/wormholecrypto",
            "email": "info@wormhole.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/governor/enqueued_vaas/": {
            "get": {
                "description": "Returns enqueued VAAs for each blockchain.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "governor-enqueued-vaas",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "Sort results in ascending or descending order.",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-array_governor_EnqueuedVaas"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/v1/governor/available_notional_by_chain": {
            "get": {
                "description": "Get available notional by chainID\nSince from the deltaswap-explorer point of view it is not a node, but has the information of all nodes,\nin order to build the endpoints it was assumed:\nThere are N number of remainingAvailableNotional values in the GovernorConfig collection. N = number of phylaxs\nfor a chainID. The smallest remainingAvailableNotional value for a chainID is used for the endpoint response.",
                "tags": [
                    "Phylax"
                ],
                "operationId": "governor-available-notional-by-chain",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/governor.AvailableNotionalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/governor/enqueued_vaas": {
            "get": {
                "description": "Get enqueued VAAs",
                "tags": [
                    "Phylax"
                ],
                "operationId": "phylaxs-enqueued-vaas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/governor.EnqueuedVaaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/governor/is_vaa_enqueued/{chain_id}/{emitter}/{seq}": {
            "get": {
                "description": "Check if vaa is enqueued",
                "tags": [
                    "Phylax"
                ],
                "operationId": "phylaxs-is-vaa-enqueued",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the blockchain",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "address of the emitter",
                        "name": "emitter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "sequence of the vaa",
                        "name": "seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/governor.EnqueuedVaaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/governor/token_list": {
            "get": {
                "description": "Get token list\nSince from the deltaswap-explorer point of view it is not a node, but has the information of all nodes,\nin order to build the endpoints it was assumed:\nFor tokens with the same originChainId and originAddress and different price values for each node,\nthe price that has most occurrences in all the nodes for an originChainId and originAddress is returned.",
                "tags": [
                    "Phylax"
                ],
                "operationId": "phylaxs-token-list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/governor.TokenList"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/phylaxset/current": {
            "get": {
                "description": "Get current phylax set.",
                "tags": [
                    "Phylax"
                ],
                "operationId": "phylax-set",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/phylax.PhylaxSetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/heartbeats": {
            "get": {
                "description": "Get heartbeats for phylaxs",
                "tags": [
                    "Phylax"
                ],
                "operationId": "phylaxs-hearbeats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/heartbeats.HeartbeatsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_deltaswapio_deltaswap-explorer_api_routes_phylax_phylax.PhylaxSet": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "index": {
                    "type": "integer"
                }
            }
        },
        "governor.AvailableNotionalItemResponse": {
            "type": "object",
            "properties": {
                "bigTransactionSize": {
                    "type": "string"
                },
                "chainId": {
                    "$ref": "#/definitions/vaa.ChainID"
                },
                "notionalLimit": {
                    "type": "string"
                },
                "remainingAvailableNotional": {
                    "type": "string"
                }
            }
        },
        "governor.AvailableNotionalResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.AvailableNotionalItemResponse"
                    }
                }
            }
        },
...
        "governor.EnqueuedVaa": {
            "type": "object",
            "properties": {
                "chainId": {
                    "$ref": "#/definitions/vaa.ChainID"
                },
                "emitterAddress": {
                    "type": "string"
                },
                "notionalValue": {
                    "type": "integer"
                },
                "sequence": {
                    "type": "string"
                },
                "txHash": {
                    "type": "string"
                }
            }
        },
...
        "governor.EnqueuedVaaItemResponse": {
            "type": "object",
            "properties": {
                "emitterAddress": {
                    "type": "string"
                },
                "emitterChain": {
                    "$ref": "#/definitions/vaa.ChainID"
                },
                "notionalValue": {
                    "type": "string"
                },
                "releaseTime": {
                    "type": "integer"
                },
                "sequence": {
                    "type": "integer"
                },
                "txHash": {
                    "type": "string"
                }
            }
        },
        "governor.EnqueuedVaaResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.EnqueuedVaaItemResponse"
                    }
                }
            }
        },
        "governor.EnqueuedVaas": {
            "type": "object",
            "properties": {
                "chainId": {
                    "$ref": "#/definitions/vaa.ChainID"
                },
                "enqueuedVaas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.EnqueuedVaa"
                    }
                }
            }
        },
...
        "governor.TokenList": {
            "type": "object",
            "properties": {
                "originAddress": {
                    "type": "string"
                },
                "originChainId": {
                    "$ref": "#/definitions/vaa.ChainID"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "phylax.PhylaxSetResponse": {
            "type": "object",
            "properties": {
                "phylaxSet": {
                    "$ref": "#/definitions/github_com_deltaswapio_deltaswap-explorer_api_routes_phylax_phylax.PhylaxSet"
                }
            }
        },
        "heartbeats.HeartbeatNetworkResponse": {
            "type": "object",
            "properties": {
                "contractAddress": {
                    "type": "string"
                },
                "errorCount": {
                    "type": "string"
                },
                "height": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "heartbeats.HeartbeatResponse": {
            "type": "object",
            "properties": {
                "p2pNodeAddr": {
                    "type": "string"
                },
                "rawHeartbeat": {
                    "$ref": "#/definitions/heartbeats.RawHeartbeat"
                },
                "verifiedPhylaxAddr": {
                    "type": "string"
                }
            }
        },
        "heartbeats.HeartbeatsResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/heartbeats.HeartbeatResponse"
                    }
                }
            }
        },
        "heartbeats.RawHeartbeat": {
            "type": "object",
            "properties": {
                "bootTimestamp": {
                    "type": "string"
                },
                "counter": {
                    "type": "string"
                },
                "features": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "phylaxAddr": {
                    "type": "string"
                },
                "networks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/heartbeats.HeartbeatNetworkResponse"
                    }
                },
                "nodeName": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "response.Response-array_governor_EnqueuedVaas": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/governor.EnqueuedVaas"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
...
        "response.ResponsePagination": {
            "type": "object",
            "properties": {
                "next": {
                    "type": "string"
                }
            }
        },
...
        "vaa.ChainID": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15,
                16,
                17,
                18,
                19,
                21,
                22,
                23,
                24,
                26,
                28,
                29,
                30,
                32,
                3104,
                10002
            ],
            "x-enum-varnames": [
                "ChainIDUnset",
                "ChainIDSolana",
                "ChainIDEthereum",
                "ChainIDTerra",
                "ChainIDBSC",
                "ChainIDPolygon",
                "ChainIDAvalanche",
                "ChainIDOasis",
                "ChainIDAlgorand",
                "ChainIDAurora",
                "ChainIDFantom",
                "ChainIDKarura",
                "ChainIDAcala",
                "ChainIDKlaytn",
                "ChainIDCelo",
                "ChainIDNear",
                "ChainIDMoonbeam",
                "ChainIDNeon",
                "ChainIDTerra2",
                "ChainIDInjective",
                "ChainIDSui",
                "ChainIDAptos",
                "ChainIDArbitrum",
                "ChainIDOptimism",
                "ChainIDPythNet",
                "ChainIDXpla",
                "ChainIDBtc",
                "ChainIDBase",
                "ChainIDSei",
                "ChainIDDeltachain",
                "ChainIDSepolia"
            ]
        },
...
    }
}